requirements.txt
langchain
openai
pathspec
python-dotenv
streamlit
tiktoken
faiss-cpu
unstructured
